Remove unneeded patch from Python-3.5.2 formula.
The patch that is failing to apply is no longer required. It was
[applied upstream](https://github.com/python/cpython/commit/6e0963c9c7849a0cbce2b715c96fbebca687e030)
(the only difference between the [patch in this
formula](https://bugs.python.org/file40478/xcode-stubs.diff) and the
upstream commit is that the upstream commit also added some comments
explaining it).

The formula seems to build fine for me by removing that patch.